Move permissions SQL Query to core_sql module
