+++ core_sql/src/modules/connection.rs
@@ -3,6 +3,7 @@ use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
 
 pub async fn start_connection() -> Pool<Postgres> {
     dotenv().ok();
+
     let url: String = std::env::var("DATABASE_URL").expect("DATABASE_URL must be set");
     PgPoolOptions::new()
         .max_connections(10)
