+++ core_sql/src/modules/permissoes/mod.rs
@@ -1,0 +1,1 @@
+pub mod select;
