+++ core_sql/src/structs/permissoes/mod.rs
@@ -1,0 +1,1 @@
+pub mod struct_permissions;
